clean architecture refactor folder structure
