+++ lib/data/datasource/remote/product_datasource_impl.dart
@@ -2,9 +2,8 @@ import 'dart:convert';
 
 import 'package:http/http.dart' as http;
 
-
 import '../../../domain/model/product_model.dart';
-import '../../../util/constant_util.dart';
+import '../../../presentation/util/constant_util.dart';
 import 'product_datasource.dart';
 
 class ProductDataSourceImpl implements ProductDataSource {
@@ -13,14 +12,14 @@ class ProductDataSourceImpl implements ProductDataSource {
   ProductDataSourceImpl({required this.client});
 
   @override
-  Future<List<ProductModel>> getProductList() async{
+  Future<List<ProductModel>> getProductList() async {
     try {
       var response = await client.get(Uri.parse('${ConstantUtil.baseUrl}products'));
       final responseBody = jsonDecode(response.body);
       List<ProductModel> products = (responseBody as List)
           .map(
             (product) => ProductModel.fromJson(product),
-      )
+          )
           .toList();
       return products;
     } catch (e) {
