+++ lib/presentation/bloc/product/product_bloc.dart
@@ -2,7 +2,7 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:store_app/domain/usecase/product_use_case.dart';
 
-import '../../domain/model/product_model.dart';
+import '../../../domain/model/product_model.dart';
 
 part 'product_event.dart';
