+++ lib/presentation/bloc/product/product_state.dart
