+++ lib/presentation/main.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:store_app/data/datasource/remote/product_datasource_impl.dart';
-import 'package:store_app/util/constant_util.dart';
 import 'package:http/http.dart' as http;
 
 import 'route_generator.dart';
+import 'util/constant_util.dart';
 
 void main() {
   runApp(const StoreApp());
