+++ lib/presentation/ui/screen/dashboard_screen.dart
@@ -5,9 +5,8 @@ import 'package:store_app/data/datasource/remote/product_datasource_impl.dart';
 import 'package:store_app/data/repository/product_repository_impl.dart';
 import 'package:store_app/domain/usecase/product_use_case.dart';
 
-import 'package:store_app/util/constant_util.dart';
-
 import '../../bloc/product/product_bloc.dart';
+import '../../util/constant_util.dart';
 import '../widget/product_card_widget.dart';
 
 class DashboardScreen extends StatelessWidget {
